% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{climateR0-package}
\alias{climateR0}
\alias{climateR0-package}
\title{climateR0: Estimate Temperature-dependent R0 for Climate-Sensitive Pathogens}
\description{
Estimate temperature dependent R0 for a given pathogen and vector from temperature data using experimentally derived constant temperature-R0 curves.
}
\author{
\strong{Maintainer}: Emilie Finch \email{emilie.finch1@lshtm.ac.uk} (\href{https://orcid.org/0000-0002-8225-7533}{ORCID})

Authors:
\itemize{
  \item Adam Kucharski \email{adam.kucharski@lshtm.ac.uk} (\href{https://orcid.org/0000-0001-8814-9421}{ORCID})
}

}
\keyword{internal}
