% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_utils.R
\name{define_prediction_starts}
\alias{define_prediction_starts}
\title{Define prediction start dates}
\usage{
define_prediction_starts(prediction_dates, horizon, n_start)
}
\arguments{
\item{prediction_dates}{Vector of dates in the dataset}

\item{horizon}{Forecast horizon in weeks}

\item{n_start}{Number of start dates to choose}
}
\value{
Vector of prediction start dates
}
\description{
Define prediction start dates
}
