% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{find_closest_days}
\alias{find_closest_days}
\title{Find closest matching day of the year}
\usage{
find_closest_days(vec1, vec2)
}
\arguments{
\item{vec1}{Vector of dates to match}

\item{vec2}{Vector of reference dates}
}
\value{
Vector of indices in vec2 that are closest to each date in vec1
}
\description{
Find closest matching day of the year
}
