% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{generate_prediction}
\alias{generate_prediction}
\title{Generate predictions using GAM-based forecasting}
\usage{
generate_prediction(
  model_data,
  input_data,
  prediction_start,
  horizon,
  pop,
  rep_prop,
  n_sim = 1,
  model_in = NULL,
  fix_all_covariates = FALSE
)
}
\arguments{
\item{model_data}{Processed model data}

\item{input_data}{Original input data}

\item{prediction_start}{Start date for prediction}

\item{horizon}{Forecast horizon in weeks}

\item{pop}{Population size}

\item{rep_prop}{Reported fraction of cases}

\item{n_sim}{Number of simulations}

\item{model_in}{Optional pre-fitted GAM model}

\item{fix_all_covariates}{Whether to fix all covariates (SIR-like model)}
}
\value{
A list containing:
\item{prediction_out}{Array of predictions}
\item{model_out}{Fitted GAM model}
}
\description{
Generate predictions using GAM-based forecasting
}
