% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_utils.R
\name{predict_cases}
\alias{predict_cases}
\title{Predict cases using a fitted GAM model}
\usage{
predict_cases(
  train_data,
  train_test_data,
  prediction_bootstrap,
  gam_mod,
  pop,
  rep_prop,
  n_sim = 100,
  fix_all_covariates
)
}
\arguments{
\item{train_data}{Training data}

\item{train_test_data}{Combined training and test data}

\item{prediction_bootstrap}{Array for storing predictions}

\item{gam_mod}{Fitted GAM model}

\item{pop}{Population size}

\item{rep_prop}{Reported fraction of cases}

\item{n_sim}{Number of simulations}

\item{fix_all_covariates}{Whether to fix all covariates (SIR-like model)}
}
\value{
Array of predictions
}
\description{
Predict cases using a fitted GAM model
}
