% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare data for GAM-based forecasting}
\usage{
prepare_data(input_data, pop, rep_prop)
}
\arguments{
\item{input_data}{A data frame containing date, cases and av_temp columns}

\item{pop}{Assumed population size}

\item{rep_prop}{Assumed proportion of infections reported as cases}
}
\value{
A list containing:
\item{model_data}{Processed data frame with additional columns for modeling}
\item{prediction_dates}{Vector of dates for prediction}
}
\description{
Prepare data for GAM-based forecasting
}
