% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tahiti_2013}
\alias{tahiti_2013}
\title{Tahiti 2014.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 25 rows and 6 columns.
}
\usage{
tahiti_2013
}
\description{
Dataset includes weekly laboratory confirmed Zika cases from Tahiti in
French Polynesia, during an outbreak in 2013 and 2014 as well as monthly
average temperature data and rainfall for Tahiti during this period
}
\details{
#' A data frame with 25 rows and 6 columns:
\describe{
\item{date}{Date of epidemiological week}
\item{cases}{Number of laboratory confirmed cases in Tahiti}
\item{reporting_sites}{Number of sentinel sites in Tahiti reporting that
month}
\item{total_sites}{Total sentinel sites Tahiti}
\item{av_temp}{Average monthly temperature for Tahiti, in °C}
\item{av_temp}{Average monthly rainfall for Tahiti, in mm}
...
}
#' @source \url{https://journals.plos.org/plosntds/article?id=10.1371/journal.pntd.0004726}
}
\keyword{datasets}
