% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{weight_cases}
\alias{weight_cases}
\title{Weight past cases by dengue serial interval}
\usage{
weight_cases(x)
}
\arguments{
\item{x}{Vector of case counts}
}
\value{
Vector of weighted case counts based on serial interval
}
\description{
Weight past cases by dengue serial interval
}
