% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature_r0.R
\name{temperature_r0}
\alias{temperature_r0}
\title{Estimate relative R0 from mean temperature}
\usage{
temperature_r0(data, vector_pathogen)
}
\arguments{
\item{data}{a numeric vector of mean temperature values in °C}

\item{vector_pathogen}{character vector specifying the vector of interest
This can be one of the following:
AeaeDENV; AeaeZIKV; AealDENV; AetaRVFV; AetaSINV; AetrEEEV; AngaPfal; CxanMVEx; CxanRRVx; CxpiSINV;
CxpiWNVxl; CxtaWEEV; CxtaWNVx; CxunWNVx}
}
\value{
numeric vector of relative R0 values corresponding to input temperature values
}
\description{
This function takes an input series of mean temperature values in °C
and returns estimated relative R0 values according to temperature-R0 curves in
from Mordecai et al 2019 https://doi.org/10.1111/ele.13335
}
