% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature_r0.R
\name{temperature_r0}
\alias{temperature_r0}
\title{Estimate relative R0 from mean temperature}
\usage{
temperature_r0(data, vector_pathogen)
}
\arguments{
\item{data}{A numeric vector of mean temperature values in °C}

\item{vector_pathogen}{Character vector specifying the vector and pathogen of interest
This can be one of the following:
AeaeDENV; AeaeZIKV; AealDENV; AetaRVFV; AetaSINV; AetrEEEV; AngaPfal; CxanMVEx; CxanRRVx; CxpiSINV;
CxpiWNVxl; CxtaWEEV; CxtaWNVx; CxunWNVx}
}
\value{
Numeric vector of relative R0 values corresponding to input temperature values
}
\description{
This function takes an input series of mean temperature values in °C
and returns estimated relative R0 values for a specified vector and pathogen.
}
\examples{
# Return relative R0 for Aedeas aegypti and dengue virus for mean temperatures from 18-36°C
temperature_r0(c(18,20,22,24,26,28,30,32,34,36), "AeaeDENV")

}
\references{
Mordecai, E. A., Caldwell, J. M., Grossman, M. K., Lippi, C. A., Johnson, L. R., Neira,
M., Rohr, J. R., Ryan, S. J., Savage, V., Shocket, M. S., Sippy, R., Stewart Ibarra, A. M.,
Thomas, M. B., Villena O. (2019). Thermal biology of mosquito-borne disease. Ecol Lett. 22(10)
\doi{10.1111/ele.13335}
}
